* main.c (main): Call cmdchk().
* text_utils.c: Include stdlib.h, string.h, styles.h and lcd.h. Add the custom_temp
and style_set extern variables to have access to them and to set them.
(cmd2int): New function to convert the second part of
the received command in to an integer. It can be a defined style or a custom
temperature.
(cmdcpy): New function to copy the first part of the received commmand into a new
string, to be analized.
(cmdchk): Check the reveived command and parse it. Set variables according to the
cmd.